requirements.txt
django-hijack>=2.1.5
